requirements/development.txt
-r base.txt

factory_boy
pytest-django
pytest-factoryboy
ipdb
ipython
